refactor: Move `RoomEvents` handling to `rooms.ts`
