+++ src/components/SmartActionBar.tsx
