+++ src/containers/ChatContainer/ChatContainer.tsx
@@ -7,7 +7,7 @@ import {twMerge} from "tailwind-merge"
 import Button, {ButtonVariant} from "../../components/Button"
 import UnreadIndicator from "../../components/UnreadIndicator"
 import Modal, {ModalRenderLocation} from "../../components/Modal"
-import SmartActionBar from "@/containers/SmartActionBar"
+import SmartActionBar from "@/components/SmartActionBar"
 import useChatInput from "./useChatInput"
 import ChatHeader from "./ChatHeader"
 import ChatInput from "./ChatInput"
