+++ src/containers/SidebarActions/DirectMessageModal.tsx
@@ -4,17 +4,12 @@ import IconButton from "../../components/IconButton"
 import {IoCheckmark, IoCloseCircle, IoCopyOutline} from "react-icons/io5"
 import Input from "../../components/Input"
 import UserProfile from "../../components/UserProfile"
-import {
-  getPartnerUserIdFromRoomDirect,
-  normalizeName,
-  stringToColor,
-  timeFormatter,
-} from "@/utils/util"
+import {normalizeName, stringToColor, timeFormatter} from "@/utils/util"
 import useConnection from "@/hooks/matrix/useConnection"
 import useInvitationLink from "@/hooks/matrix/useInvitationLink"
 import useUsersSearch from "@/hooks/matrix/useUserSearch"
-import {useSidebarModalActiveStore} from "./useSidebarActions"
-import {getDirectRoomsIds} from "@/utils/rooms"
+import {useSidebarModalActiveStore} from "./hooks/useSidebarActions"
+import {getDirectRoomsIds, getPartnerUserIdFromRoomDirect} from "@/utils/rooms"
 
 type DirectChatRecentProps = {
   userId: string
