+++ src/containers/SidebarActions/NotificationsModal.tsx
@@ -8,12 +8,12 @@ import Button, {
 import Typography, {TypographyVariant} from "../../components/Typography"
 import {twMerge} from "tailwind-merge"
 import IconButton from "../../components/IconButton"
-import {useSidebarModalActiveStore} from "./useSidebarActions"
+import {useSidebarModalActiveStore} from "./hooks/useSidebarActions"
 import useCachedNotifications, {
   useNotificationsStateStore,
   type LocalNotificationData,
-} from "./useCachedNotifications"
-import {markAllNotificationsAsRead} from "@/utils/notifications"
+} from "./hooks/useCachedNotifications"
+import {markAllNotificationsAsRead} from "@/containers/SidebarActions/hooks/notifications"
 import Notification from "./Notification"
 
 export type NotificationActions = {
