+++ src/containers/SidebarActions/hooks/notifications.ts
@@ -1,4 +1,4 @@
-import {type LocalNotificationData} from "@/containers/SidebarActions/useCachedNotifications"
+import {type LocalNotificationData} from "@/containers/SidebarActions/hooks/useCachedNotifications"
 
 const NOTIFICATIONS_LOCAL_STORAGE_KEY = "notifications"
