+++ src/containers/SidebarActions/hooks/useCachedNotifications.ts
@@ -1,4 +1,4 @@
-import {getNotificationsData} from "@/utils/notifications"
+import {getNotificationsData} from "@/containers/SidebarActions/hooks/notifications"
 import {useEffect, useState} from "react"
 import {create} from "zustand"
