+++ src/containers/SidebarActions/hooks/useSidebarActions.ts
@@ -1,4 +1,4 @@
-import useConnection from "../../hooks/matrix/useConnection"
+import useConnection from "../../../hooks/matrix/useConnection"
 import {useNavigate} from "react-router-dom"
 import {ViewPath} from "@/utils/util"
 import {useCallback} from "react"
