+++ src/hooks/matrix/useRooms.ts
@@ -6,7 +6,7 @@ import {normalizeName} from "@/utils/util"
 import {RoomType, type RoomProps} from "@/components/Room"
 import {useActiveSpaceIdStore} from "./useSpaces"
 import useActiveRoomIdStore from "@/hooks/matrix/useActiveRoomIdStore"
-import {saveNotification} from "@/utils/notifications"
+import {saveNotification} from "@/containers/SidebarActions/hooks/notifications"
 import {getNotificationFromInviteEvent} from "./useGlobalEventListeners"
 import {KnownMembership} from "matrix-js-sdk/lib/@types/membership"
 import {getDirectRoomsIds, getRoomsFromSpace} from "@/utils/rooms"
