+++ src/utils/rooms.ts
@@ -4,29 +4,26 @@ import {
   type MatrixClient,
   Direction,
   EventType,
+  HistoryVisibility,
+  JoinRule,
+  type MatrixEvent,
+  MsgType,
+  type User,
 } from "matrix-js-sdk"
 import {
   assert,
   CommonAssertion,
+  deleteMessage,
   getImageUrl,
-  ImageSizes,
   normalizeName,
   stringToColor,
 } from "./util"
 import {type RosterUserProps, UserPowerLevel} from "@/components/RosterUser"
 import {UserStatus} from "@/components/UserProfile"
-import {getRoomAdminsAndModerators} from "./members"
-
-export function getDirectRoomsIds(client: MatrixClient): string[] {
-  const directRooms = client.getAccountData("m.direct")
-  const content = directRooms?.event.content
-
-  if (content === undefined) {
-    return []
-  }
-
-  return Object.values(content).flat()
-}
+import {getRoomAdminsAndModerators, isUserRoomAdmin} from "./members"
+import {type AnyMessage, MessageKind} from "@/hooks/matrix/useActiveRoom"
+import {KnownMembership} from "matrix-js-sdk/lib/@types/membership"
+import {type MessageBaseProps} from "@/components/MessageContainer"
 
 export function getRoomsFromSpace(
   spaceId: string,
@@ -81,24 +78,10 @@ export function getRoomsFromSpace(
   return storeRooms
 }
 
-export function isDirectRoom(client: MatrixClient | null, room: Room): boolean {
-  if (client === null) {
-    return false
-  }
-
-  const myUserId = client.getUserId()
-
-  assert(myUserId !== null, CommonAssertion.UserIdNotFound)
-
-  return (
-    room
-      .getLiveTimeline()
-      .getState(Direction.Forward)
-      ?.events.get(EventType.RoomMember)
-      // Find event by userId.
-      // If the client user is not in the room event then this room is not a direct chat for the user.
-      ?.get(myUserId)?.event.content?.is_direct ?? false
-  )
+export enum ImageSizes {
+  Server = 47,
+  MessageAndProfile = 40,
+  ProfileLarge = 60,
 }
 
 export async function getRoomMembers(
@@ -181,4 +164,645 @@ export async function getRoomMembers(
   }
 
   return membersProperty
+}
+
+// #region Direct rooms
+
+export function getDirectRoomsIds(client: MatrixClient): string[] {
+  const directRooms = client.getAccountData("m.direct")
+  const content = directRooms?.event.content
+
+  if (content === undefined) {
+    return []
+  }
+
+  return Object.values(content).flat()
+}
+
+export function isDirectRoom(client: MatrixClient | null, room: Room): boolean {
+  if (client === null) {
+    return false
+  }
+
+  const myUserId = client.getUserId()
+
+  assert(myUserId !== null, CommonAssertion.UserIdNotFound)
+
+  return (
+    room
+      .getLiveTimeline()
+      .getState(Direction.Forward)
+      ?.events.get(EventType.RoomMember)
+      // Find event by userId.
+      // If the client user is not in the room event then this room is not a direct chat for the user.
+      ?.get(myUserId)?.event.content?.is_direct ?? false
+  )
+}
+
+export function getPartnerUserIdFromRoomDirect(room: Room): string {
+  assert(
+    room.getJoinedMemberCount() === 2,
+    "Direct chat must have exactly two participants."
+  )
+
+  const userId = room
+    .getJoinedMembers()
+    .find(member => member.userId !== room.myUserId)?.userId
+
+  assert(
+    userId !== undefined,
+    "If one participant is the current user, the other must exist."
+  )
+
+  return userId
+}
+
+// #region Events
+
+export const handleRoomEvents = async (
+  client: MatrixClient,
+  activeRoom: Room
+): Promise<AnyMessage[]> => {
+  const roomHistory = await client.scrollback(activeRoom, 30)
+  const events = roomHistory.getLiveTimeline().getEvents()
+  const lastReadEventId = getLastReadEventIdFromRoom(activeRoom, client)
+  const isAdminOrModerator = isUserRoomAdmin(activeRoom, client)
+  const allMessageProperties: AnyMessage[] = []
+
+  for (let index = 0; index < events.length; index++) {
+    const event = events[index]
+
+    const messageProperties = await handleEvents(
+      client,
+      event,
+      roomHistory.roomId,
+      isAdminOrModerator
+    )
+
+    if (messageProperties === null) {
+      continue
+    }
+
+    allMessageProperties.push(messageProperties)
+
+    if (lastReadEventId === event.getId() && index !== events.length - 1) {
+      allMessageProperties.push({
+        kind: MessageKind.Unread,
+        data: {lastReadEventId},
+      })
+    }
+
+    void client.sendReadReceipt(event)
+  }
+
+  return allMessageProperties
+}
+
+export const handleEvents = async (
+  client: MatrixClient,
+  event: MatrixEvent,
+  roomId: string,
+  isAdminOrModerator: boolean
+): Promise<AnyMessage | null> => {
+  const timestamp = event.localTimestamp
+  const sender = event.sender?.userId ?? null
+
+  if (sender === null) {
+    return null
+  }
+
+  const user = client.getUser(sender)?.displayName
+
+  if (user === undefined) {
+    return null
+  }
+
+  switch (event.getType()) {
+    case EventType.RoomMessage: {
+      return await handleMessagesEvent(
+        client,
+        timestamp,
+        event,
+        roomId,
+        sender,
+        isAdminOrModerator
+      )
+    }
+    case EventType.RoomMember: {
+      return handleMemberEvent(user, timestamp, client, event)
+    }
+    case EventType.RoomTopic: {
+      return await handleRoomTopicEvent(user, timestamp, event)
+    }
+    case EventType.RoomGuestAccess: {
+      return await handleGuestAccessEvent(user, timestamp, event)
+    }
+    case EventType.RoomHistoryVisibility: {
+      return await handleHistoryVisibilityEvent(user, timestamp, event)
+    }
+    case EventType.RoomJoinRules: {
+      return await handleJoinRulesEvent(user, timestamp, event)
+    }
+    case EventType.RoomCanonicalAlias: {
+      return await handleRoomCanonicalAliasEvent(user, timestamp, event)
+    }
+    case EventType.RoomAvatar: {
+      return await handleRoomAvatarEvent(user, timestamp, event)
+    }
+    case EventType.RoomName: {
+      return await handleRoomNameEvent(user, timestamp, event)
+    }
+    default: {
+      return null
+    }
+  }
+}
+
+export function getLastReadEventIdFromRoom(
+  room: Room,
+  client: MatrixClient
+): string | null {
+  const userId = client.getUserId()
+
+  if (userId === null) {
+    return null
+  }
+
+  const eventReadUpTo = room.getEventReadUpTo(userId)
+
+  if (eventReadUpTo === null) {
+    return null
+  }
+
+  return room.findEventById(eventReadUpTo)?.getId() ?? null
+}
+
+export function handleMemberLeave(
+  user: string,
+  displayName?: string,
+  previousDisplayName?: string,
+  previousMembership?: string
+): string | null {
+  if (
+    displayName === undefined ||
+    previousDisplayName === undefined ||
+    previousMembership === undefined
+  ) {
+    return null
+  }
+
+  switch (previousMembership) {
+    case KnownMembership.Invite: {
+      return `${user} has canceled the invitation to ${previousDisplayName}`
+    }
+    case KnownMembership.Ban: {
+      return `${user} has removed the ban from ${displayName}`
+    }
+    case KnownMembership.Join: {
+      return `${user} has left the room`
+    }
+    default: {
+      return null
+    }
+  }
+}
+
+export const handleMemberEvent = (
+  user: string,
+  timestamp: number,
+  client: MatrixClient,
+  event: MatrixEvent
+): AnyMessage | null => {
+  const previousContent = event.getUnsigned().prev_content
+  const eventContent = event.getContent()
+  const stateKey = event.getStateKey()
+  const eventId = event.event.event_id
+  const displayName = eventContent.displayname
+  const previousMembership = previousContent?.membership
+  const previousDisplayName = previousContent?.displayname
+  let text: string | null = null
+
+  if (stateKey === undefined || eventId === undefined) {
+    return null
+  }
+
+  switch (eventContent.membership) {
+    // TODO: Handle here other types of RoomMember events.
+    case KnownMembership.Join: {
+      if (
+        previousMembership === KnownMembership.Invite ||
+        previousMembership !== KnownMembership.Join
+      ) {
+        text = `${user} has joined to the room`
+      } else if (
+        previousDisplayName !== undefined &&
+        displayName !== previousDisplayName
+      ) {
+        text = `${previousDisplayName} has change the name to ${displayName}`
+      } else if (
+        eventContent.avatar_url !== undefined &&
+        previousContent?.avatar_url === undefined
+      ) {
+        text = `${displayName} has put a profile photo`
+      } else if (
+        eventContent.avatar_url !== undefined &&
+        eventContent.avatar_url !== previousContent?.avatar_url
+      ) {
+        text = `${displayName} has change to the profile photo`
+      } else if (
+        eventContent.avatar_url === undefined &&
+        previousContent?.avatar_url !== undefined
+      ) {
+        text = `${displayName} has remove the profile photo`
+      }
+
+      break
+    }
+    case KnownMembership.Invite: {
+      text = `${user} invited ${displayName}`
+
+      break
+    }
+    case KnownMembership.Ban: {
+      text = `${user} has banned ${previousDisplayName}: ${eventContent.reason}`
+
+      break
+    }
+    case KnownMembership.Leave: {
+      text = handleMemberLeave(
+        user,
+        client.getUser(stateKey)?.displayName,
+        previousDisplayName,
+        previousMembership
+      )
+
+      break
+    }
+    case undefined: {
+      return null
+    }
+  }
+
+  if (text === null) {
+    // If event is not handled or the event has error.
+    return null
+  }
+
+  return {
+    kind: MessageKind.Event,
+    data: {text, timestamp, id: eventId},
+  }
+}
+
+export const handleGuestAccessEvent = async (
+  user: string,
+  timestamp: number,
+  event: MatrixEvent
+): Promise<AnyMessage | null> => {
+  let text: string | null = null
+  const guestAccess = event.getContent().guest_access
+
+  switch (guestAccess) {
+    case "can_join": {
+      text = `${user} authorized anyone to join the room`
+
+      break
+    }
+    case "forbidden": {
+      text = `${user} has prohibited guests from joining the room`
+
+      break
+    }
+    case "restricted": {
+      text = `${user} restricted guest access to the room. Only guests with valid tokens can join.`
+
+      break
+    }
+    case "knock": {
+      text = `${user} enabled "knocking" for guests. Guests must request access to join.`
+
+      break
+    }
+    default: {
+      console.warn("Unknown guest access type:", guestAccess)
+    }
+  }
+
+  const eventId = event.event.event_id
+
+  if (text === null || eventId === undefined) {
+    return null
+  }
+
+  return {
+    kind: MessageKind.Event,
+    data: {
+      text,
+      timestamp,
+      id: eventId,
+    },
+  }
+}
+
+export const handleJoinRulesEvent = async (
+  user: string,
+  timestamp: number,
+  event: MatrixEvent
+): Promise<AnyMessage | null> => {
+  let text: string | null = null
+
+  switch (event.getContent().join_rule) {
+    case JoinRule.Invite: {
+      text = `${user} restricted the room to guests`
+
+      break
+    }
+    case JoinRule.Public: {
+      text = `${user} made the room public to anyone who knows the link.`
+
+      break
+    }
+    case JoinRule.Restricted: {
+      text = `${user} made the room private. Only admins can invite now.`
+
+      break
+    }
+    default: {
+      console.warn("Unknown join rule:", event.getContent().join_rule)
+    }
+  }
+
+  const eventId = event.event.event_id
+
+  if (text === null || eventId === undefined) {
+    return null
+  }
+
+  return {
+    kind: MessageKind.Event,
+    data: {
+      text,
+      timestamp,
+      id: eventId,
+    },
+  }
+}
+
+export const handleRoomTopicEvent = async (
+  user: string,
+  timestamp: number,
+  event: MatrixEvent
+): Promise<AnyMessage | null> => {
+  const topic = event.getContent().topic
+  const eventId = event.event.event_id
+
+  if (eventId === undefined) {
+    return null
+  }
+
+  const text =
+    topic === undefined
+      ? `${user} has remove the topic of the room`
+      : `${user} has change to the topic to <<${topic}>>`
+
+  return {
+    kind: MessageKind.Event,
+    data: {
+      text,
+      timestamp,
+      id: eventId,
+    },
+  }
+}
+
+export const handleHistoryVisibilityEvent = async (
+  user: string,
+  timestamp: number,
+  event: MatrixEvent
+): Promise<AnyMessage | null> => {
+  let content: string | null = null
+
+  switch (event.getContent().history_visibility) {
+    case HistoryVisibility.Shared: {
+      content = `${user} made the future history of the room visible to all members of the room`
+
+      break
+    }
+    case HistoryVisibility.Invited: {
+      content = `${user} made the room future history visible to all room members, from the moment they are invited.`
+
+      break
+    }
+    case HistoryVisibility.Joined: {
+      content = `${user} made the room future history visible to all room members, from the moment they are joined.`
+
+      break
+    }
+    case HistoryVisibility.WorldReadable: {
+      content = `${user} made the future history of the room visible to anyone people.`
+
+      break
+    }
+  }
+
+  const eventId = event.event.event_id
+
+  if (content === null || eventId === undefined) {
+    return null
+  }
+
+  return {
+    kind: MessageKind.Event,
+    data: {
+      text: content,
+      timestamp,
+      id: eventId,
+    },
+  }
+}
+
+export const handleRoomCanonicalAliasEvent = async (
+  user: string,
+  timestamp: number,
+  event: MatrixEvent
+): Promise<AnyMessage | null> => {
+  const alias = event.getContent().alias
+
+  const text =
+    alias === undefined
+      ? `${user} has remove the main address for this room`
+      : `${user} set the main address for this room as ${alias}`
+
+  const eventId = event.event.event_id
+
+  if (eventId === undefined) {
+    return null
+  }
+
+  return {
+    kind: MessageKind.Event,
+    data: {
+      text,
+      timestamp,
+      id: eventId,
+    },
+  }
+}
+
+export const handleRoomAvatarEvent = async (
+  user: string,
+  timestamp: number,
+  event: MatrixEvent
+): Promise<AnyMessage | null> => {
+  const eventId = event.event.event_id
+
+  if (eventId === undefined) {
+    return null
+  }
+
+  const text =
+    event.getContent().url === undefined
+      ? `${user} has remove the avatar for this room`
+      : `${user} changed the avatar of the room`
+
+  return {
+    kind: MessageKind.Event,
+    data: {
+      text,
+      timestamp,
+      id: eventId,
+    },
+  }
+}
+
+export const handleRoomNameEvent = async (
+  user: string,
+  timestamp: number,
+  event: MatrixEvent
+): Promise<AnyMessage | null> => {
+  const text = `${user} has changed the room name to ${event.getContent().name}`
+  const eventId = event.event.event_id
+
+  if (eventId === undefined) {
+    return null
+  }
+
+  return {
+    kind: MessageKind.Event,
+    data: {
+      text,
+      timestamp,
+      id: eventId,
+    },
+  }
+}
+
+export const handleMessagesEvent = async (
+  client: MatrixClient,
+  timestamp: number,
+  event: MatrixEvent,
+  roomId: string,
+  sender: string,
+  isAdmin: boolean
+): Promise<AnyMessage | null> => {
+  const eventContent = event.getContent()
+  const user = client.getUser(sender)
+  const isMessageOfMyUser = client.getUserId() === sender
+
+  if (user === null) {
+    return null
+  }
+
+  const eventId = event.event.event_id
+  const authorDisplayName = user.displayName ?? user.userId
+
+  if (eventId === undefined) {
+    return null
+  }
+
+  const messageBaseProperties: MessageBaseProps = {
+    authorAvatarUrl: getImageUrl(
+      user.avatarUrl,
+      client,
+      ImageSizes.MessageAndProfile
+    ),
+    authorDisplayName,
+    authorDisplayNameColor: stringToColor(authorDisplayName),
+    id: eventId,
+    onAuthorClick: () => {},
+    text: eventContent.body,
+    timestamp,
+    onDeleteMessage:
+      isMessageOfMyUser || isAdmin
+        ? () => {
+            deleteMessage(client, roomId, eventId)
+          }
+        : undefined,
+  }
+
+  switch (eventContent.msgtype) {
+    case MsgType.Text: {
+      return {
+        kind: MessageKind.Text,
+        data: {...messageBaseProperties},
+      }
+    }
+    case MsgType.Image: {
+      return {
+        kind: MessageKind.Image,
+        data: {
+          ...messageBaseProperties,
+          text: "",
+          imageUrl: getImageUrl(eventContent.url as string, client),
+        },
+      }
+    }
+    case undefined: {
+      return convertToMessageDeletedProperties(client, user, timestamp, event)
+    }
+    default: {
+      return null
+    }
+  }
+}
+
+const convertToMessageDeletedProperties = (
+  client: MatrixClient,
+  user: User,
+  timestamp: number,
+  event: MatrixEvent
+): AnyMessage | null => {
+  const eventId = event.event.event_id
+  const deletedBy = event.getUnsigned().redacted_because?.sender
+  const authorDisplayName = user.displayName ?? user.userId
+
+  if (eventId === undefined || deletedBy === undefined) {
+    return null
+  }
+
+  const deletedByUser = client.getUser(deletedBy)?.displayName
+
+  if (deletedByUser === undefined) {
+    return null
+  }
+
+  const reason = event.getUnsigned().redacted_because?.content.reason
+
+  const text =
+    reason === undefined
+      ? `${deletedByUser} has delete this message`
+      : `${deletedByUser} has delete this message because <<${reason}>>`
+
+  return {
+    kind: MessageKind.Text,
+    data: {
+      authorAvatarUrl: getImageUrl(user.avatarUrl, client),
+      authorDisplayName,
+      authorDisplayNameColor: stringToColor(authorDisplayName),
+      onAuthorClick: () => {},
+      text,
+      timestamp,
+      id: eventId,
+    },
+  }
 }
